+++ CHANGELOG.md
@@ -1,5 +1,29 @@
 # Changelog
 
+## `v0.0.60`
+
+### Features
+
+- Parse `EthermintAccount` in `toBaseAccount` util function
+
+## `v0.0.59`
+
+### Fixes
+
+- Fix Station mobile wallet to not throw errors if pub keys do not exist
+
+## `v0.0.58`
+
+### Fixes
+
+- Fix Station mobile wallet connect issue when connecting more than 1 chain
+
+## `v0.0.57`
+
+### Fixes
+
+- Allow Terra testnet to work correctly with Station extension
+
 ## `v0.0.56`
 
 ### Features